+++ src/Exception/CiNotDetectedException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace OndraM\CiDetector\Exception;
+
+class CiNotDetectedException extends \RuntimeException
+{
+}