+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There is always Unreleased section on the top. Subsections (Added, Changed, Fixed, Removed) should be added as needed. -->
 
 ## Unreleased
+- BC: The `detect()` method of `CiDetector` class is no longer static.
 
 ## 1.1.0 - 2017-01-06
 - Add `getGitBranch()` method to detect Git branch of the build (supported by all CIs except TeamCity).