+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ## Unreleased
 - BC: The `detect()` method of `CiDetector` class is no longer static.
 - BC: Rearrange namespaces, all classes are now in `CiDetector` sub-namespace.
+- Added `isCiDetected()` method to detect if current environment is CI.
+- BC: `detect()` method always returns instance of `CiInterface` and throws `CiNotDetectedException` if CI environment is not detected.
 
 ## 1.1.0 - 2017-01-06
 - Add `getGitBranch()` method to detect Git branch of the build (supported by all CIs except TeamCity).